+++ hermes-management/src/main/java/pl/allegro/tech/hermes/management/config/CorsConfiguration.java
@@ -1,0 +1,10 @@
+package pl.allegro.tech.hermes.management.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties(CorsProperties.class)
+public class CorsConfiguration {
+
+}
