+++ hermes-management/src/main/java/pl/allegro/tech/hermes/management/config/SupportTeamConfiguration.java
@@ -1,0 +1,27 @@
+package pl.allegro.tech.hermes.management.config;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.ClientHttpRequestFactory;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+@EnableConfigurationProperties(SupportTeamServiceProperties.class)
+public class SupportTeamConfiguration {
+
+    @Bean("managementRequestFactory")
+    @ConfigurationProperties(prefix = "management.restTemplate")
+    public ClientHttpRequestFactory clientHttpRequestFactory() {
+        return new HttpComponentsClientHttpRequestFactory();
+    }
+
+    @Bean
+    public RestTemplate restTemplate(@Qualifier("managementRequestFactory") ClientHttpRequestFactory clientHttpRequestFactory) {
+        return new RestTemplate(clientHttpRequestFactory);
+    }
+
+}
